+++ src/extension/Create Procedure/Procedure Creator/CreateProcedureAL0118.ts
@@ -19,9 +19,11 @@ export class CreateProcedureAL0118 implements ICreateProcedure {
     syntaxTree: SyntaxTree | undefined;
     document: TextDocument;
     diagnostic: Diagnostic;
+    private returnType: { analyzed: boolean, type: string | undefined }
     constructor(document: TextDocument, diagnostic: Diagnostic) {
         this.document = document;
         this.diagnostic = diagnostic;
+        this.returnType = { analyzed: false, type: undefined }
     }
     async initialize() {
         this.syntaxTree = await SyntaxTree.getInstance(this.document);
@@ -55,13 +57,16 @@ export class CreateProcedureAL0118 implements ICreateProcedure {
         return false;
     }
     async getReturnType(): Promise<string | undefined> {
+        if(this.returnType.analyzed)
+            return this.returnType.type
         if (!this.syntaxTree) { return undefined; }
         let invocationExpressionTreeNode: ALFullSyntaxTreeNode | undefined = this.syntaxTree.findTreeNode(this.diagnostic.range.start, [FullSyntaxTreeNodeKind.getInvocationExpression()]) as ALFullSyntaxTreeNode;
         let invocationExpressionRange: Range = TextRangeExt.createVSCodeRange(invocationExpressionTreeNode.fullSpan);
         invocationExpressionRange = DocumentUtils.trimRange(this.document, invocationExpressionRange);
 
-        let returnType: string | undefined = await TypeDetective.findReturnTypeOfInvocationAtPosition(this.document, invocationExpressionRange.start);
-        return returnType;
+        this.returnType.type = await TypeDetective.findReturnTypeOfInvocationAtPosition(this.document, invocationExpressionRange.start);
+        this.returnType.analyzed = true
+        return this.returnType.type;
     }
     async getObject(): Promise<ALObject> {
         let objectTreeNode: ALFullSyntaxTreeNode | undefined = SyntaxTreeExt.getObjectTreeNode(this.syntaxTree as SyntaxTree, this.diagnostic.range.start);
@@ -76,11 +81,14 @@ export class CreateProcedureAL0118 implements ICreateProcedure {
     containsSnippet(): boolean {
         return false;
     }
-    isReturnTypeRequired(): boolean {
+    async isReturnTypeRequired(): Promise<boolean> {
         let invocationNode: ALFullSyntaxTreeNode | undefined = this.syntaxTree?.findTreeNode(this.diagnostic.range.start, [FullSyntaxTreeNodeKind.getInvocationExpression()]);
-        if (invocationNode)
+        if (invocationNode) {
             if (invocationNode.parentNode!.kind == FullSyntaxTreeNodeKind.getPageField())
                 return true
+            if (invocationNode.parentNode!.kind == FullSyntaxTreeNodeKind.getArgumentList() && (await this.getReturnType()) === undefined)
+                return true
+        }
         return false
     }
 }