+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to the "al-codeactions" extension will be documented in this
   - Create procedure in argument list even if the procedure does not exist yet
   - Create Procedure in Validate Statement
   - Create overload of procedure: fix if the new parameter is a text variable
+  - Remove unused variables: Show locations of unused parameters which were not removed to be able to fix them manually.
 
 ## 1.0.2